+++ phy/gui/__init__.py
@@ -6,6 +6,8 @@
 from .qt import start_qt_app, run_qt_app, qt_app, enable_qt
 from .dock import DockWindow
 
-from .base import (BaseViewModel, HTMLViewModel, WidgetCreator,
-                   BaseGUI, BaseSession,
+from .base import (BaseViewModel,
+                   HTMLViewModel,
+                   WidgetCreator,
+                   BaseGUI,
                    )
