+++ phy/gui/base.py
@@ -9,15 +9,10 @@
 
 from collections import Counter
 import inspect
-import os.path as op
 
 from ..ext.six import string_types
 from ..utils._misc import _show_shortcuts
 from ..utils import debug, EventEmitter
-from ..utils.settings import (Settings,
-                              _ensure_dir_exists,
-                              _phy_user_dir,
-                              )
 from ._utils import _wrap_html, _read
 from .dock import DockWindow
 
@@ -553,178 +548,3 @@ class BaseGUI(EventEmitter):
     def exit(self):
         """Close the GUI."""
         self.close()
-
-
-#------------------------------------------------------------------------------
-# Session
-#------------------------------------------------------------------------------
-
-class BaseSession(EventEmitter):
-    """Give access to the data, views, and GUIs in an interactive session.
-
-    The model must implement:
-
-    * `model(path)`
-    * `model.path`
-    * `model.close()`
-
-    Events
-    ------
-
-    open
-    close
-
-    """
-    def __init__(self,
-                 model=None,
-                 path=None,
-                 phy_user_dir=None,
-                 default_settings_path=None,
-                 vm_classes=None,
-                 gui_classes=None,
-                 ):
-        super(BaseSession, self).__init__()
-
-        self.model = None
-        if phy_user_dir is None:
-            phy_user_dir = _phy_user_dir()
-        _ensure_dir_exists(phy_user_dir)
-        self.phy_user_dir = phy_user_dir
-
-        self._create_settings(default_settings_path)
-
-        self._view_creator = WidgetCreator(widget_classes=vm_classes)
-
-        if gui_classes is None:
-            gui_classes = self.settings['gui_classes']
-        self._gui_creator = WidgetCreator(widget_classes=gui_classes)
-
-        self.connect(self.on_open)
-        self._pre_open()
-        if model or path:
-            self.open(path, model=model)
-
-    def _create_settings(self, default_settings_path):
-        self.settings = Settings(phy_user_dir=self.phy_user_dir,
-                                 default_path=default_settings_path,
-                                 )
-
-        @self.connect
-        def on_open():
-            # Initialize the settings with the model's path.
-            self.settings.on_open(self.experiment_path)
-
-    # Methods to override
-    # -------------------------------------------------------------------------
-
-    def _pre_open(self):
-        pass
-
-    def _create_model(self, path):
-        """Create a model from a path.
-
-        Must be overriden.
-
-        """
-        pass
-
-    def _save_model(self):
-        """Save a model.
-
-        Must be overriden.
-
-        """
-        pass
-
-    def on_open(self):
-        pass
-
-    # File-related actions
-    # -------------------------------------------------------------------------
-
-    def open(self, path=None, model=None):
-        """Open a dataset."""
-        # Close the session if it is already open.
-        if self.model:
-            self.close()
-        if model is None:
-            model = self._create_model(path)
-        self.model = model
-        self.experiment_path = (op.realpath(path)
-                                if path else self.phy_user_dir)
-        self.emit('open')
-
-    def reopen(self):
-        self.open(model=self.model)
-
-    def save(self):
-        self._save_model()
-
-    def close(self):
-        """Close the currently-open dataset."""
-        self.model.close()
-        self.emit('close')
-        self.model = None
-
-    # Views and GUIs
-    # -------------------------------------------------------------------------
-
-    def show_gui(self, name=None, show=True, **kwargs):
-        """Show a new GUI."""
-        if name is None:
-            gui_classes = list(self._gui_creator.widget_classes.keys())
-            if gui_classes:
-                name = gui_classes[0]
-
-        # Get the default GUI config.
-        params = {p: self.settings.get('{}_{}'.format(name, p), None)
-                  for p in ('config', 'shortcuts', 'state')}
-        params.update(kwargs)
-
-        # Create the GUI.
-        gui = self._gui_creator.add(name,
-                                    model=self.model,
-                                    settings=self.settings,
-                                    **params)
-        gui._save_state = True
-
-        # Connect the 'open' event.
-        self.connect(gui.on_open)
-
-        @gui.connect
-        def on_close_gui():
-            self.unconnect(gui.on_open)
-            # Save the params of every view in the GUI.
-            for vm in gui.views:
-                self.save_view_params(vm, save_size_pos=False)
-            gs = gui.main_window.save_geometry_state()
-            gs['view_count'] = gui.view_count()
-            if not gui._save_state:
-                gs['state'] = None
-                gs['geometry'] = None
-            self.settings['{}_state'.format(name)] = gs
-            self.settings.save()
-
-        # HACK: do not save GUI state when views have been closed or reset
-        # in the session, otherwise Qt messes things up in the GUI.
-        @gui.connect
-        def on_close_view(view):
-            gui._save_state = False
-
-        @gui.connect
-        def on_reset_gui():
-            gui._save_state = False
-
-        if show:
-            gui.show()
-
-        return gui
-
-    def save_view_params(self, vm, save_size_pos=True):
-        """Save the parameters exported by a view model instance."""
-        to_save = vm.exported_params(save_size_pos=save_size_pos)
-        for key, value in to_save.items():
-            assert vm.name
-            name = '{}_{}'.format(vm.name, key)
-            self.settings[name] = value
-            debug("Save {0}={1} for {2}.".format(name, value, vm.name))
