+++ phy/io/__init__.py
@@ -4,5 +4,6 @@
 """Input/output."""
 
 from .h5 import File, open_h5
+from .base_model import BaseModel, BaseSession
 from .kwik.model import KwikModel
 from .store import ClusterStore, StoreItem
